refactor(scores): move Exporter to dspy-o11y-langfuse gem
Split score reporting code for cleaner gem separation:
- Core DSPy gem: DSPy.score(), ScoreEvent, DataType, Evaluators
- dspy-o11y-langfuse: ScoresExporter (Langfuse REST API client)

The ScoresExporter is now at:
DSPy::Observability::Adapters::Langfuse::ScoresExporter

Bump versions:
- dspy: 0.33.0 → 0.34.0
- dspy-o11y-langfuse: 1.0.1 → 1.1.0
