+++ lib/dspy/o11y/langfuse/version.rb
@@ -3,7 +3,7 @@
 module DSPy
   module O11y
     module Langfuse
-      VERSION = '1.0.1'
+      VERSION = '1.1.0'
     end
   end
 end
