+++ lib/dspy/scores.rb
@@ -2,7 +2,6 @@
 
 require_relative 'scores/data_type'
 require_relative 'scores/score_event'
-require_relative 'scores/exporter'
 require_relative 'scores/evaluators'
 
 module DSPy
