+++ lib/dspy/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module DSPy
-  VERSION = "0.33.0"
+  VERSION = "0.34.0"
 end
